backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-filter>=24.2
celery>=5.4
redis>=5.0
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=22.0
